Initial egrid facility to FRS matching script
Goal is to get FRS IDs for all egrid plants. First step is getting matches from the FRS bridge file. The next steps is using a file with CAMD and FRS matches provided by the eGRID team.  For non-matches, the next step is to begin location and names based matching with the master FRS national facility file. The goal is to produce a master table of ORIS Plant IDs and FRS IDs that can then be used to match across other databases.