Total_gen file created, name corrected before
There is Nitrous oxide and Nitrous Oxide. (Look at the red colored data) So this becomes two different flows. The elementary flow name matcher corrects this afterwards, so now we see the name has matched.

So the solution is that we should correct the flow names before. Not in the dictionary process writer but in the generation_process_builder_fnc.

Hence I am shifting that.