update UUIDs and compartments from coal mining; addresses #285
Note that the elementary flows in the 2023 coal mining inventory are not mapped to FEDEFL, particularly the ones with the NETL source. Manually mapped all new UUIDs and compartments to be consistent with the rest of the code. Could this be solved by just sending coal_df through concat_map_upstream_databases in get_upstream_process_df?